Fix #390 KIT inframe insertion annotation issue
- When hgvsp is missing from VEP output and hgvsc contains "dup", use
the entire amino_acids field instead of just the first one.
- Unrelated: it seems that we sometimes return hgvspShort without the p.
prefix. I believe that is an error, so fixed those as well.
- Because of reuse of the mock classes, mockito complains about mocking
certain methods that aren't used. In this case I would say reusing the
initialized mocks is useful but not all examples are useful for testing
all resolvers. So I added the "Silent" runner, we have already been
using this in some places after upgrading to the latest mockito. We
might want to revisit this at a later point, but seemed out of the scope
for this PR