Retain original query input in annotation JSON response
This commit addresses an issue where the original input query did not
match the 'input' field in the JSON annotation response due to a shift in
the value of 'start_position' for certain types of genomic variants when
converting genomic locations to HGVSg format strings.

As a result some variants submitted for annotation through the Genome
Nexus annotation pipeline were considered failed annotations since the
input query did not match the 'variant' field exactly in the JSON
response which is how the client matches the queries back to the records
submitted for annotation.

Signed-off-by: Angelica Ochoa <15623749+ao508@users.noreply.github.com>