verification of variant Reference_Allele string
Principal changes
- textual description/explaination appears below this summary
- HGVS and genomicLocation annotation queries are examined
- annotation objects from underlying VEP system contain reference
    alleles, and these are used for comparison/validation
- if full reference allele is not present, a follow-up query is made
    to obtain it
- annotation fails if user-provided reference allele differs from VEP
- utilities written to extract reference allele from Annotations
- unit tests for Verified AnnotationService classes added:
    - HGVS query format unit tests
    - genomicLocation query format unit tests
Additional changes
- business logic removed from AnnotationController : selection of
    AnnotationService is made in a upper-tier selection service class
- all variant annotation services now set originalVariantQuery
- an integration test of long deletions has been disabled with a
    comment in the code. The query format was invalid and a related
    bug has been reported as an issue.
Unresolved Questions
- what formats should be accepted by the /annotation endpoint? We are
    primarily supporting HGVS format here, but it appears that other
    formats were at one time considered valid. If this is HGVS format
    only we can simplify code.

When reference allele specifications are present in variant queries, those
sequences are now checked against the reference genome sequence present in
the underlying VEP cache. If the specified sequence does not match the
reference assembly sequence in use by the prediction software, the variant
annotation fails.

There are two variant formats accepted by the genome-nexus api: HGVS and
genomicLocation. The HGVS standard now recommends that reference allele
specification be omitted for all types except for substitutions (where it
is still required). However, the Ensembl VEP service API still allows HGVS
queries which are fully specified with reference allele, so genome-nexus
also allows it. When reference alleles are specified (in delins, del,
substitutions) they are checked for match to the reference assembly.
For genomicLocation queries, the genomic location always includes a
specified reference allele. For genomicLocation variants of type
substitution, delins, and deletion, the specified reference sequence is
checked against the reference genome.

genome-nexus can be configured to use the public Ensembl server VEP api as
an underlying annotation system, or can use a locally deployed VEP command
line tool (repo genome-nexus-vep) as the underlying annotation system. When
using the public Esembl server, queries are converted to HGVS. When using
the local VEP deployment, queries are converte to REST-like Region format.
In both systems, we usually receive Annotation results which include both
the full reference allele and the full variant allele sequences. These
strings are compared to the query string for validation. However, because of
the different formats / services we are using in the underlying system we
sometimes receive a shortened version of the reference allele. In these
(rare) cases a subsequent query is performed to obtain the full reference
assembly sequence for comparison.

Co-authored-by: Avery Wang <18199796+averyniceday@users.noreply.github.com>
Co-authored-by: Robert Sheridan <7747489+sheridancbio@users.noreply.github.com>